Feature/playlist

+++ src/WebApp/Application/Services/SongService.cs
@@ -1,4 +1,6 @@
-﻿using Application.DTO.Songs;
+﻿using Application.Authorization;
+using Application.DTO.Songs;
+using Application.Errors;
 using Application.InfrastructureInterfaces;
 using Application.Interfaces;
 using AutoMapper;
@@ -14,11 +16,13 @@ internal class SongService : ISongService
     private readonly IUnitOfWork _uow;
     private readonly IMapper _mapper;
     private readonly IFileHandler _fileHandler;
-    public SongService(IUnitOfWork uow, IMapper mapper, IFileHandler fileHandler)
+    private readonly ISongAuthorizationService _songAuthorizationService;
+    public SongService(IUnitOfWork uow, IMapper mapper, IFileHandler fileHandler, ISongAuthorizationService songAuthorizationService)
     {
         _uow = uow;
         _mapper = mapper;
         _fileHandler = fileHandler;
+        _songAuthorizationService = songAuthorizationService;
     }
 
     public async Task<Result<SongResponse>> CreateSongAsync(CreateSongRequest model, string userId, CancellationToken cancellationToken = default)
@@ -27,14 +31,14 @@ internal class SongService : ISongService
 
         if (exists)
         {
-            return Result.Fail($"Song with name {model.Name} arleady exists");
+            return new ValidationError($"Song with name {model.Name} already exists");
         }
 
         var allExist = model.GenreIds.TrueForAll(id => _uow.Exists<Genre>(e => e.GenreId == id));
 
         if (!allExist)
         {
-            return Result.Fail($"Please provide valid genres, one of provided does not exist");
+            return new NotFoundError($"One or many of provided genres do not exists");
         }
 
         var song = _mapper.Map<Song>(model);
@@ -50,13 +54,13 @@ internal class SongService : ISongService
         return Result.Ok(_mapper.Map<SongResponse>(song));
     }
 
-    public async Task<Result<SongResponse>> GetSongAsync(int songId, CancellationToken cancellationToken = default)
+    public async Task<Result<SongResponse>> GetSongAsync(int songId, string? userId, CancellationToken cancellationToken = default)
     {
-        var song = await _uow.SongRepository.GetByIdAsync(songId, false, cancellationToken);
+        var song = await _uow.SongRepository.GetByIdAsync(songId, false, userId, cancellationToken);
 
         if (song is null)
         {
-            return Result.Fail($"Song with Id - {songId} doesn't exist");
+            return new NotFoundError($"Song with id {songId} cannot be found");
         }
 
         return Result.Ok(_mapper.Map<SongResponse>(song));
@@ -71,18 +75,30 @@ internal class SongService : ISongService
 
     public async Task<Result<SongResponse>> UpdateSongAsync(UpdateSongRequest model, CancellationToken cancellationToken = default)
     {
-        var nameExists = await _uow.ExistsAsync<Song>(x => x.Name.ToUpper() == model.Name.ToUpper(), cancellationToken);
-
-        if (nameExists)
-        {
-            return Result.Fail($"Song with name {model.Name} already exists");
-        }
-
-        var song = await _uow.SongRepository.GetByIdAsync(model.SongId, true, cancellationToken);
+        var song = await _uow.SongRepository.GetByIdAsync(model.SongId, cancellationToken: cancellationToken);
 
         if (song is null)
         {
-            return Result.Fail($"Song with Id - {model.SongId} doesn't exist");
+            return new NotFoundError($"Song with id {model.SongId} cannot be found");
+        }
+
+        if (!_songAuthorizationService.Authorize(song, OperationType.Update))
+        {
+            return new ForbiddenAccessError();
+        }
+
+        var nameExists = await _uow.ExistsAsync<Song>(x => x.Name.ToUpper() == model.Name.ToUpper() && x.SongId != model.SongId, cancellationToken);
+
+        if (nameExists)
+        {
+            return new NotFoundError($"Song with name {model.Name} already exists");
+        }
+
+        var allExist = model.GenreIds.TrueForAll(id => _uow.Exists<Genre>(e => e.GenreId == id));
+
+        if (!allExist)
+        {
+            return new NotFoundError($"One or many of provided genres do not exists");
         }
 
         _mapper.Map(model, song);
@@ -97,14 +113,20 @@ internal class SongService : ISongService
 
     public async Task<Result> DeleteSongAsync(int songId, CancellationToken cancellationToken = default)
     {
-        var songUrl = await _uow.SongRepository.DeleteAsync(songId, cancellationToken);
+        var song = await _uow.SongRepository.GetByIdAsync(songId, false, cancellationToken: cancellationToken);
 
-        if (songUrl is null)
+        if (song is null)
         {
-            return Result.Fail($"Song with Id - {songId} doesn't exist");
+            return new NotFoundError($"Song with id {songId} cannot be found");
         }
 
-        var fileName = Path.GetFileName(songUrl);
+        if (!_songAuthorizationService.Authorize(song, OperationType.Delete))
+        {
+            return new ForbiddenAccessError();
+        }
+
+        await _uow.SongRepository.DeleteAsync(songId, cancellationToken);
+        var fileName = Path.GetFileName(song.Url);
 
         await _fileHandler.DeleteFileAsync(fileName, FileContainer.Songs, cancellationToken);
         await _uow.SaveChangesAsync(cancellationToken);
@@ -114,18 +136,23 @@ internal class SongService : ISongService
 
     public async Task<Result<SongResponse>> UploadPhotoAsync(int songId, IFormFile photo, CancellationToken cancellationToken = default)
     {
-        var song = await _uow.SongRepository.GetByIdAsync(songId, true, cancellationToken);
+        var song = await _uow.SongRepository.GetByIdAsync(songId, cancellationToken: cancellationToken);
 
         if (song is null)
         {
-            return Result.Fail($"Song with Id - {songId} doesn't exist");
+            return new NotFoundError($"Song with id {songId} cannot be found");
+        }
+
+        if (!_songAuthorizationService.Authorize(song, OperationType.Patch))
+        {
+            return new ForbiddenAccessError();
         }
 
         var photoUrl = song.Photo is null 
             ? await _fileHandler.UploadFileAsync(photo, FileContainer.Photos, cancellationToken)
-            : await _fileHandler.UpdateFileAsync(Path.GetFileName(song.Url), photo, FileContainer.Photos, cancellationToken);
+            : await _fileHandler.UpdateFileAsync(Path.GetFileName(song.Photo.URL), photo, FileContainer.Photos, cancellationToken);
 
-        Photo songPhoto = new Photo()
+        var songPhoto = new Photo()
         {
             URL = photoUrl,
             Size = photo.Length,
@@ -133,30 +160,35 @@ internal class SongService : ISongService
         };
 
         song.Photo = songPhoto;
-        await _uow.SaveChangesAsync(cancellationToken);
+        await _uow.SaveChangesAsync();
 
         return Result.Ok(_mapper.Map<SongResponse>(song));
     }
 
     public async Task<Result> DeletePhotoAsync(int songId, CancellationToken cancellationToken = default)
     {
-        var song = await _uow.SongRepository.GetByIdAsync(songId, true, cancellationToken);
+        var song = await _uow.SongRepository.GetByIdAsync(songId, cancellationToken: cancellationToken);
 
         if (song is null)
         {
-            return Result.Fail($"Song with Id - {songId} doesn't exist");
+            return new NotFoundError($"Song with id {songId} cannot be found");
+        }
+
+        if (!_songAuthorizationService.Authorize(song, OperationType.Delete))
+        {
+            return new ForbiddenAccessError();
         }
 
         if (song.Photo is null)
         {
-            return Result.Fail($"There is not photo to be deleted");
+            return Result.Ok();
         }
 
         var fileName = Path.GetFileName(song.Photo.URL);
         await _fileHandler.DeleteFileAsync(fileName, FileContainer.Photos, cancellationToken);
 
         await _uow.PhotoRepository.DeleteAsync(song.Photo.PhotoId);
-        await _uow.SaveChangesAsync(cancellationToken);
+        await _uow.SaveChangesAsync();
 
         return Result.Ok();
     }
